+++ src/ThreadPool.cpp
@@ -11,7 +11,7 @@ ThreadPool::ThreadPool( size_t size, void *(*func) (void *), void *args ) {
     pthread_attr_t attr;
     pthread_attr_init(&attr);
     pthread_attr_setdetachstate(&attr, PTHREAD_CREATE_JOINABLE);
-    mlog(INT_DAPI, "THREAD_POOL: Creating %d threads", size ); 
+    mlog(INT_DAPI, "THREAD_POOL: Creating %lu threads", (unsigned long)size ); 
     for( size_t t = 0; t < size; t++ ) {
         if ( 0 != pthread_create(&threads[t], &attr, func, args) ) {
             thread_error = errno;
